requirements.txt
pillow==10.4.0
PyAutoGui==0.9.54
opencv-python==4.10.0.84
numpy==2.1.1